index: split RevWalk types to "rev_walk" module
Added pub(super) where needed.
